% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{hclust}
\alias{hclust}
\title{Hierarchical clustering}
\usage{
hclust(d, method = "complete", members = NULL)
}
\arguments{
\item{d}{a dissimilarity structure as produced by \code{dist}.}

\item{method}{the agglomeration method to be used.  This should
    be (an unambiguous abbreviation of) one of
    \code{"ward.D"}, \code{"ward.D2"}, \code{"single"}, \code{"complete"},
    \code{"average"} (= UPGMA), \code{"mcquitty"} (= WPGMA),
    \code{"median"} (= WPGMC) or \code{"centroid"} (= UPGMC).}

\item{members}{\code{NULL} or a vector with length size of
    \code{d}. See the \sQuote{Details} section.}
}
\description{
Experiments on hierarchical clustering
}
\seealso{
stats::hclust
}
